# docs/reference/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ArtifactHistoryEntry": {
        "additionalProperties": true,
        "description": "One persisted artifact snapshot discovered in session message history.",
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/TripPlanArtifact"
          },
          "message_index": {
            "title": "Message Index",
            "type": "integer"
          },
          "message_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Timestamp"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "artifact",
          "message_index"
        ],
        "title": "ArtifactHistoryEntry",
        "type": "object"
      },
      "ArtifactHistoryResponse": {
        "additionalProperties": true,
        "description": "HTTP response payload for retrieving persisted artifact history for one session.",
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ArtifactHistoryEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ArtifactHistoryResponse",
        "type": "object"
      },
      "Attraction": {
        "description": "Simplified attraction payload exposed by city endpoints.",
        "properties": {
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "duration": {
            "title": "Duration",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "ticket": {
            "title": "Ticket",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "duration",
          "ticket"
        ],
        "title": "Attraction",
        "type": "object"
      },
      "BudgetReportArtifact": {
        "additionalProperties": true,
        "description": "Budget and runtime-cost summary attached to the artifact payload.",
        "properties": {
          "executionBudget": {
            "additionalProperties": true,
            "title": "Executionbudget",
            "type": "object"
          },
          "fallbackSteps": {
            "default": 0,
            "title": "Fallbacksteps",
            "type": "integer"
          },
          "staleResultCount": {
            "default": 0,
            "title": "Staleresultcount",
            "type": "integer"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "title": "BudgetReportArtifact",
        "type": "object"
      },
...
      "CityAttractionsResponse": {
        "description": "Response schema for a city's attraction recommendations.",
        "properties": {
          "attractions": {
            "items": {
              "$ref": "#/components/schemas/Attraction"
            },
            "title": "Attractions",
            "type": "array"
          },
          "city": {
            "title": "City",
            "type": "string"
          }
        },
        "required": [
          "city",
          "attractions"
        ],
        "title": "CityAttractionsResponse",
        "type": "object"
      },
      "CityDetail": {
        "description": "Expanded city detail payload with recommendation metadata.",
        "properties": {
          "attractions": {
            "items": {
              "$ref": "#/components/schemas/Attraction"
            },
            "title": "Attractions",
            "type": "array"
          },
          "avg_budget_per_day": {
            "title": "Avg Budget Per Day",
            "type": "integer"
          },
          "best_seasons": {
            "items": {
              "type": "string"
            },
            "title": "Best Seasons",
            "type": "array"
          },
          "data_source": {
            "const": "curated",
            "title": "Data Source",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "editorial_note": {
            "title": "Editorial Note",
            "type": "string"
          },
          "family_friendly": {
            "title": "Family Friendly",
            "type": "boolean"
          },
          "food_friendly": {
            "title": "Food Friendly",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rain_friendly": {
            "title": "Rain Friendly",
            "type": "boolean"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "style_label": {
            "title": "Style Label",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trip_duration": {
            "title": "Trip Duration",
            "type": "string"
          },
          "walk_intensity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Walk Intensity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "region",
          "tags",
          "description",
          "avg_budget_per_day",
          "best_seasons",
          "trip_duration",
          "walk_intensity",
          "rain_friendly",
          "family_friendly",
          "food_friendly",
          "style_label",
          "editorial_note",
          "data_source",
          "attractions"
        ],
        "title": "CityDetail",
        "type": "object"
      },
      "CityListResponse": {
        "description": "Response schema for city list endpoint.",
        "properties": {
          "cities": {
            "items": {
              "$ref": "#/components/schemas/CitySummary"
            },
            "title": "Cities",
            "type": "array"
          }
        },
        "required": [
          "cities"
        ],
        "title": "CityListResponse",
        "type": "object"
      },
      "CitySummary": {
        "description": "Compact city card used in list responses.",
        "properties": {
          "avg_budget_per_day": {
            "title": "Avg Budget Per Day",
            "type": "integer"
          },
          "best_seasons": {
            "items": {
              "type": "string"
            },
            "title": "Best Seasons",
            "type": "array"
          },
          "data_source": {
            "const": "curated",
            "title": "Data Source",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "editorial_note": {
            "title": "Editorial Note",
            "type": "string"
          },
          "family_friendly": {
            "title": "Family Friendly",
            "type": "boolean"
          },
          "food_friendly": {
            "title": "Food Friendly",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rain_friendly": {
            "title": "Rain Friendly",
            "type": "boolean"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "style_label": {
            "title": "Style Label",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trip_duration": {
            "title": "Trip Duration",
            "type": "string"
          },
          "walk_intensity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Walk Intensity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "region",
          "tags",
          "description",
          "avg_budget_per_day",
          "best_seasons",
          "trip_duration",
          "walk_intensity",
          "rain_friendly",
          "family_friendly",
          "food_friendly",
          "style_label",
          "editorial_note",
          "data_source"
        ],
        "title": "CitySummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItineraryDraftArtifact": {
        "additionalProperties": true,
        "description": "Draft plan steps and validation details for the itinerary stage.",
        "properties": {
          "explanation": {
            "default": "",
            "title": "Explanation",
            "type": "string"
          },
          "planId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planid"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Steps",
            "type": "array"
          },
          "validationErrors": {
            "items": {},
            "title": "Validationerrors",
            "type": "array"
          },
          "validationStatus": {
            "default": "pass",
            "title": "Validationstatus",
            "type": "string"
          }
        },
        "title": "ItineraryDraftArtifact",
        "type": "object"
      },
...
      "LatestArtifactResponse": {
        "additionalProperties": true,
        "description": "HTTP response payload for retrieving the latest persisted trip artifact.",
        "properties": {
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPlanArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifact_found": {
            "default": false,
            "title": "Artifact Found",
            "type": "boolean"
          },
          "message_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Index"
          },
          "message_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Timestamp"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "LatestArtifactResponse",
        "type": "object"
      },
...
      "ResearchDossierArtifact": {
        "additionalProperties": true,
        "description": "Research summary and evidence bundle surfaced to the client.",
        "properties": {
          "destinations": {
            "items": {
              "type": "string"
            },
            "title": "Destinations",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ResearchEvidenceArtifact"
            },
            "title": "Evidence",
            "type": "array"
          },
          "sourceTools": {
            "items": {
              "type": "string"
            },
            "title": "Sourcetools",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "title": "ResearchDossierArtifact",
        "type": "object"
      },
      "ResearchEvidenceArtifact": {
        "additionalProperties": true,
        "description": "One evidence item collected during research-oriented tool calls.",
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          }
        },
        "title": "ResearchEvidenceArtifact",
        "type": "object"
      },
...
      "TripIntentArtifact": {
        "additionalProperties": true,
        "description": "Structured intent summary captured for one trip-planning run.",
        "properties": {
          "confidence": {
            "title": "Confidence"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "entities": {
            "additionalProperties": true,
            "title": "Entities",
            "type": "object"
          },
          "name": {
            "default": "general",
            "title": "Name",
            "type": "string"
          }
        },
        "title": "TripIntentArtifact",
        "type": "object"
      },
      "TripPlanArtifact": {
        "additionalProperties": true,
        "description": "Top-level artifact envelope shared across stream and history payloads.",
        "properties": {
          "answer": {
            "default": "",
            "title": "Answer",
            "type": "string"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetReportArtifact"
          },
          "intent": {
            "$ref": "#/components/schemas/TripIntentArtifact"
          },
          "itinerary": {
            "$ref": "#/components/schemas/ItineraryDraftArtifact"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "reasoning": {
            "default": "",
            "title": "Reasoning",
            "type": "string"
          },
          "research": {
            "$ref": "#/components/schemas/ResearchDossierArtifact"
          },
          "toolsUsed": {
            "items": {
              "type": "string"
            },
            "title": "Toolsused",
            "type": "array"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationReportArtifact"
          }
        },
        "title": "TripPlanArtifact",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationReportArtifact": {
        "additionalProperties": true,
        "description": "Verification outcome, retry hints, and issue list for one run.",
        "properties": {
          "issues": {
            "items": {},
            "title": "Issues",
            "type": "array"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          },
          "refreshTargets": {
            "items": {
              "type": "string"
            },
            "title": "Refreshtargets",
            "type": "array"
          },
          "shouldRetry": {
            "default": false,
            "title": "Shouldretry",
            "type": "boolean"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "title": "VerificationReportArtifact",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI Travel Assistant API with SSE streaming support.",
    "title": "moyuan-travel-agent API",
    "version": "3.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/artifacts/{session_id}/history": {
      "get": {
        "description": "Return persisted artifact snapshots for one session in newest-first order.",
        "operationId": "get_artifact_history_api_artifacts__session_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact History",
        "tags": [
          "artifact"
        ]
      }
    },
    "/api/artifacts/{session_id}/latest": {
      "get": {
        "description": "Return the latest persisted trip artifact for one session.",
        "operationId": "get_latest_artifact_api_artifacts__session_id__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Artifact",
        "tags": [
          "artifact"
        ]
      }
    },
...
    "/api/cities": {
      "get": {
        "description": "List cities with optional region and tag filters.",
        "operationId": "list_cities_api_cities_get",
        "parameters": [
          {
            "description": "Filter by region",
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "minLength": 1,
                  "pattern": ".*\\S.*",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by region",
              "title": "Region"
            }
          },
          {
            "description": "Filter by comma-separated tags",
            "in": "query",
            "name": "tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "pattern": ".*\\S.*",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by comma-separated tags",
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cities",
        "tags": [
          "city"
        ]
      }
    },
    "/api/cities/{city_id}": {
      "get": {
        "description": "Get full city detail by city identifier.",
        "operationId": "get_city_api_cities__city_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "City Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get City",
        "tags": [
          "city"
        ]
      }
    },
    "/api/cities/{city_id}/attractions": {
      "get": {
        "description": "Get recommended attractions for the selected city.",
        "operationId": "get_city_attractions_api_cities__city_id__attractions_get",
        "parameters": [
          {
            "in": "path",
            "name": "city_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "City Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityAttractionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get City Attractions",
        "tags": [
          "city"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Snapshot server",
      "url": "http://localhost:38000"
    }
  ]
}